Syntax for quasiquotes
In line with Martin's SID "String interpolation and formatting", I've implemented a syntax for qq's:
https://docs.google.com/document/d/1NdxNxZYodPA-c4MLr33KzwzKFkzm9iW9POexT9PkJsU/edit?hl=en_US

The machinery described in the SID has been generalized to allow for
arbitrary quasiquotes, not only string interpolations.

The main difference from the proposal is using the newly introduced Quasiquote class from scala.reflect:
  class Quasiquote(parts: List[String], splices: List[Any], formats: List[String])
and transforming quasiquotes into an intermediate representation that involves Quasiquote and that
can be processed in a domain-specific way later on, by a quasiquoting macro.

For example,
  s"hello $s"
will be transformed into
  val parts = Apply(Ident(ListModule), List(Literal(Constant("hello ")), Literal(Constant(""))))
  val splices = Apply(Ident(ListModule), Ident(newTermName("s")))
  val formats = Apply(Ident(ListModule), List(Literal(Constant(""))))
  val ctor = New(Ident(QuasiquoteLiteralClass), List(parts, splices, formats))
  Apply(Select(ctor, newTermName("s")), List())

For now I've injected hacks into doTypedApply and typedCase that detect calls to s()
and manually assemble string interpolation logic. Later on, the typer will resolve "s"
to be an extension macro for Quasiquote and will invoke that macro, which, in its turn,
will produce the glue that involves StringContext and Formatted as mentioned in the SID.