Changed scheme to handle existentialSkolems.
Instead of an existentialSkolem pointing back to the tree that created it, we now keep it as an attachment in the tree. Advantage (1) This is stable under tree copying (2) We can now transform  existential skolems together with the trees that created them. This should fix a range of bugs, both long-standing and recent, including SI-6025